keep picker inside the window
If the button is positioned too close to the window edge, just using
'left_bottom()' of the button will result in it being partiially
clipped.

This change solves the problem by putting it there initially, saving the
size of the picker, and next redraw checking if that is outside the
bounds, and moving it inwards if so. This results potentially 1 frame
where it is actually outside the bounds before being corrected. egui
doesn't support any way to measure a widget, so there's no other way
around this.

This also requires removing the 'movable' feature, as I could not get it
to work with movable picker. In fact, movable picker seems kind of
broken anyway, if the window is resized, it doesn't move with the
button...